+++ src/main/java/blood_speed/step/data/SpeedData.java
@@ -1,11 +1,11 @@
 package blood_speed.step.data;
 
 public class SpeedData {
-    public final int speed;
+    public final double speed;
     public final int[][] matrix;
     public final int[][] pde;
 
-    public SpeedData(int speed, int[][] matrix, int[][] pde) {
+    public SpeedData(double speed, int[][] matrix, int[][] pde) {
         this.speed = speed;
         this.matrix = matrix;
         this.pde = pde;
